requirements.txt
fastapi
uvicorn
pydantic
python-multipart
numpy
Pillow